Refactor code relating to vehicle details and subtype options, fix infinite loop bug for subtype options
