+++ frontend/src/features/manageVehicles/helpers/vehicleSubtypes.ts
@@ -1,3 +1,10 @@
+import {
+  BaseVehicle,
+  PowerUnit,
+  Trailer,
+  VEHICLE_TYPES,
+} from "../types/Vehicle";
+
 /**
  * Determine whether or not a vehicle subtype ic considered to be LCV.
  * @param subtype Vehicle subtype
@@ -11,4 +18,15 @@ export const EMPTY_VEHICLE_SUBTYPE = {
   typeCode: "",
   type: "",
   description: "",
+};
+
+export const selectedVehicleSubtype = (vehicle: BaseVehicle) => {
+  switch (vehicle.vehicleType) {
+    case VEHICLE_TYPES.POWER_UNIT:
+      return (vehicle as PowerUnit).powerUnitTypeCode;
+    case VEHICLE_TYPES.TRAILER:
+      return (vehicle as Trailer).trailerTypeCode;
+    default:
+      return "";
+  }
 };
